fullstack-chat-app-complete/backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
bcrypt
SQLAlchemy
python-dotenv
scikit-learn
joblib
requests
openai
azure-storage-blob
atlassian-python-api
pytesseract
Pillow
faiss-cpu
